Commit old changes, I think adding in the ability to compute a score for each decrypt which doesn't currently work

+++ Set1/CheckDecryptUtilities.h
@@ -2,4 +2,4 @@
 
 extern const float ascii_frequencies[];
 bool byte_stream_is_ascii(unsigned char *bytes);
-int byte_stream_compare_character_distribution(unsigned char *bytes);
+float byte_stream_compare_character_distribution(unsigned char *bytes);
